Incrementing Score Implemented
The score increments per second based on a rate set by the
PlayerController. The PlayerController requires a reference to the
ScoreController to control setting the increment rate, pausing the
score increment on collision with an object, resuming the score
increment when the player begins running again, and any future controls
such as increasing the score in bursts when something is picked up like
a power.