In Scene Restart Screen
I set Time.timeScale to 0, unhide the game over UI elements, and when
someone hits the restart button it reloads the Master Scene and resets
the timeScale to 1.

I was observing what looked like UI images and text glitches towards
the end of developing this. I debugged a few values but found that the
values dictating the hp bar length were fine, contrary to the visual
representation. Ima let it be for now and fix it in the future if it
continues to be an issue.

Deleted the game over scene.