Ammo UI, Pickups, and Shoooooting
Did all of those.
Changed controls. For PC jumping is now ‘W’ while shooting is spacebar.
On a touch device jumping is tapping on the right side of the screen
while shooting is tapping on the left side. The reason why shooting is
on the left side is because I plan to have enemies slowly chase the
player from the left side of the screen. Enemies can trip on ground
obstacles just like the player, or jump over (chance roll?), but can
also be shot.