SpawnGroundObstacle Method Finished
Spawns ground obstacles in varying group sizes depending on the
difficulty (0f - 1f) and maximum group size (int). The spawnLocations
of ground obstacles are checked with previously spawned ground
obstacles, in all chunks, and if they are too close according to a
value calculated by the difficulty variable, that prospective ground
obstacle will not be spawned.